# packages/api-spec/generated/openapi.yaml
openapi: 3.1.0
info:
  title: Bytecrunch Contracts API
  version: 0.0.0
...
paths:
...
  /v1/integrations:
    get:
      operationId: IntegrationOperations_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Integration'
      security:
        - ContractsOAuth2:
            - agreements:read
            - agreements:write
            - templates:read
            - templates:write
            - webhooks:manage
    post:
      operationId: IntegrationOperations_create
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/IntegrationCredential'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIntegration'
      security:
        - ContractsOAuth2:
            - agreements:read
            - agreements:write
            - templates:read
            - templates:write
            - webhooks:manage
  /v1/integrations/{integrationKey}/rotate-secret:
    post:
      operationId: IntegrationOperations_rotateSecret
      parameters:
        - name: integrationKey
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/IntegrationCredential'
                  - $ref: '#/components/schemas/ErrorResponse'
      security:
        - ContractsOAuth2:
            - agreements:read
            - agreements:write
            - templates:read
            - templates:write
            - webhooks:manage
...
  /v1/notifications:
    get:
      operationId: NotificationOperations_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
      security:
        - ContractsOAuth2:
            - agreements:read
            - agreements:write
            - templates:read
            - templates:write
            - webhooks:manage
  /v1/notifications/read-all:
    post:
      operationId: NotificationOperations_readAll
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
                    format: int32
                required:
                  - updated
      security:
        - ContractsOAuth2:
            - agreements:read
            - agreements:write
            - templates:read
            - templates:write
            - webhooks:manage
  /v1/notifications/{notificationId}/read:
    post:
      operationId: NotificationOperations_read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Notification'
                  - $ref: '#/components/schemas/ErrorResponse'
      security:
        - ContractsOAuth2:
            - agreements:read
            - agreements:write
            - templates:read
            - templates:write
            - webhooks:manage
...
  /v1/webhooks:
    get:
      operationId: WebhookOperations_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookEndpoint'
      security:
        - ContractsOAuth2:
            - agreements:read
            - agreements:write
            - templates:read
            - templates:write
            - webhooks:manage
    post:
      operationId: WebhookOperations_create
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/WebhookEndpoint'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhook'
      security:
        - ContractsOAuth2:
            - agreements:read
            - agreements:write
            - templates:read
            - templates:write
            - webhooks:manage
components:
...
  schemas:
...
    CreateIntegration:
      type: object
      required:
        - key
        - name
        - mappingStrategy
        - allowedRedirectUris
      properties:
        key:
          type: string
        name:
          type: string
        mappingStrategy:
          type: string
          enum:
            - host_asserted
            - shared_oidc
            - account_linking
        allowedRedirectUris:
          type: array
          items:
            type: string
            format: uri
        allowedOrigins:
          type: array
          items:
            type: string
            format: uri
        identityProviderKey:
          type: string
          enum:
            - participant-oidc
        scopes:
          type: array
          items:
            type: string
            enum:
              - conditions:read
              - signing_sessions:write
...
    CreateWebhook:
      type: object
      required:
        - url
        - events
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        issues:
          type: array
          items:
            type: string
...
    Integration:
      type: object
      required:
        - id
        - tenantId
        - key
        - name
        - mappingStrategy
        - allowedRedirectUris
        - allowedOrigins
        - identityProviderKey
        - clientId
        - scopes
        - createdAt
      properties:
        id:
          type: string
        tenantId:
          type: string
        key:
          type: string
        name:
          type: string
        mappingStrategy:
          type: string
          enum:
            - host_asserted
            - shared_oidc
            - account_linking
        allowedRedirectUris:
          type: array
          items:
            type: string
            format: uri
        allowedOrigins:
          type: array
          items:
            type: string
            format: uri
        identityProviderKey:
          anyOf:
            - type: string
              enum:
                - participant-oidc
            - type: 'null'
        clientId:
          anyOf:
            - type: string
            - type: 'null'
        scopes:
          type: array
          items:
            type: string
            enum:
              - conditions:read
              - signing_sessions:write
        createdAt:
          type: string
          format: date-time
...
    IntegrationCredential:
      type: object
      required:
        - integration
        - clientSecret
      properties:
        integration:
          $ref: '#/components/schemas/Integration'
        clientSecret:
          type: string
...
    Notification:
      type: object
      required:
        - id
        - tenantId
        - recipientPersonId
        - recipientEmail
        - type
        - title
        - body
        - agreementId
        - threadId
        - actorName
        - readAt
        - createdAt
      properties:
        id:
          type: string
        tenantId:
          type: string
        recipientPersonId:
          type: string
        recipientEmail:
          type: string
        type:
          type: string
        title:
          type: string
        body:
          type: string
        agreementId:
          type: string
        threadId:
          anyOf:
            - type: string
            - type: 'null'
        actorName:
          type: string
        readAt:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        createdAt:
          type: string
          format: date-time
...
    WebhookEndpoint:
      type: object
      required:
        - id
        - url
        - events
        - createdAt
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
  securitySchemes:
    ContractsOAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: http://localhost:8080/realms/bytecrunch/protocol/openid-connect/token
          scopes:
            agreements:read: ''
            agreements:write: ''
            templates:read: ''
            templates:write: ''
            webhooks:manage: ''
    IntegrationOAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: http://localhost:3001/oauth/token
          scopes:
            conditions:read: ''
            signing_sessions:write: ''
servers:
  - url: http://localhost:3001
    description: Local development
    variables: {}